requirements.txt
streamlit==1.23.1
web3==5.31.4
protobuf==3.20.3
eth-brownie==1.19.3
slither-analyzer==0.9.6
mythril==0.23.25
pandas==2.1.1
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.21
eth-abi==2.2.0
python-dotenv==1.0.1
web3==6.11.0
markdown
z3-solver
